requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-multipart
aiofiles
sqlmodel
asyncpg
psycopg2-binary
greenlet
redis
numpy
opencv-python
openslide-python
instanseg-torch[full]
torch
requests
shapely
zarr
